project and save job some feature fixed

+++ job_finder_backend/app/Http/Controllers/ProjectController.php
@@ -71,7 +71,7 @@ class ProjectController extends Controller
             //image file store at public/image folder
             if ($request->hasFile('image')) {
                 $oldImagePath = public_path('image/' . $project->image);
-                if (file_exists($oldImagePath)) {
+                if (file_exists($oldImagePath) && is_file($oldImagePath)) {
                     unlink($oldImagePath);
                 }
 
@@ -90,9 +90,16 @@ class ProjectController extends Controller
         }
     }
 
-    public function destory($id)
+    public function destroy($id)
     {
         $project = Project::findOrFail($id);
+        //remove existing image
+        if($project->image){
+            $imagePath = public_path('image/' . $project->image);
+            if (file_exists($imagePath) && is_file($imagePath)) {
+                unlink($imagePath);
+            }
+        }
         $project->delete();
         return response()->json(['message' => 'Project deleted successfully!'], 200);
     }
